requirements.txt
pillow==10.3.0
numpy==1.26.4

unstructured==0.13.7
unstructured-client==0.22.0
unstructured-inference==0.7.31
unstructured.pytesseract==0.3.12

Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
Werkzeug==3.0.3

paddleocr==2.7.3
paddlepaddle==2.6.1
pikepdf
tabula-py==2.9.0
JPype1==1.5.0
psycopg2-binary==2.9.9
pillow_heif

